Fixes ISSUE-53776: There should not be readonly buttons
The readonly property of the buttons that had it set to true has been set to false. There is no point in
having readonly buttons. This property was being ignored, unless when there was a Role->Window Access configuration
defined for the tab that contains those buttons. When issue https://issues.openbravo.com/view.php\?id\=53408 was
fixed, this problem was made more visible