fixes BUG-53918: BO utilization causes excessive JVM metaspace usage
Backoffice uses OBBindings to compute defaults and filter expressions. They are
executed as JavaScript scripts. Each time one was used, it was compiled and
discarded causing a high usage of the mataspace.

Now they are cached (their key is the script string itself) so that each of them
is compiled just once.