[dalbootstrapmapping] Add reflist values to reference in PostLoad to avoid race condition
We cannot guarantee that the value attribute will be mapped before the reference attribute. The setReference method was taking
for granted that the value attribute would have been already mapped, but that is not the case.

To fix that, associate the reflist values with their respective reference in a PostLoad method, once all the attributes have
been mapped