fixes BUG-54099: cannot nullify password
Password fields could not be nullified from BO windows. When trying to set to
null a password reference field from a window, the request does not include
password_cleartext property, which made the change to be skipped.

Now we check whether the value is null so it is set even if there is no
cleartext property.