Fixes ISSUE-53866: 'No items to show' is shown when adding the 'Price List Version' to the new product
The 'Price List Version' dropdown list does not display any rows when creating a new product in Backoffice.

The problem was in 4 validation rules that use role_orgAccess table, so it is a regression related with recently merged automatic roles project.

That project removed all the role_orgaccess entries, as automatic roles now by default have access to all orgs of the client with their corresponding user-level.

To fix this, we added an extra query "EXISTS" before the old one to be executed where the role is automatic so the second will be executed when the role is manual.