[HU Type Sequence] Centralize sequence management logic
After these are the places that work with sequences:
1. SequenceUtil.java: Java implementation of sequences, with the new features
2. FIN_Utility.java: It calls the SequenceUtil methods
3. AD_SEQUENCE_DOCTYPE and AD_SEQUENCE_DOC: PL/SQL procedures are not modified, and do not support (yet) the
new features
4. Utility.java: It calls the above PL/SQL procedures

Created new class SequenceUtil to centralize the sequence management. This class
supports the new feature of calculation types, length type and control digit.

Moved code in charge of mananaging sequences in FIN_Utility class to SequenceUtil.
All the public methods in FIN_Utility now references to SequenceUtil.

SequenceUtil does not support the old Autonumbering = 'N', because it is going to be deprecated
in this project.
This type of sequence is a dummy sequence. When linked to a Document Type, the sequence to calculate
the document number was taken from ANOTHER sequence with name DocumentNo_<tableName>, for example
DocumentNo_C_Order. These sequences are created at Initial Client Setup (see AD_CLIENT_TRG and AD_COLUMN_TRG2)
and they are shared amoung all the organizations.

To work with Autonumbering = 'N' sequence, the developer needs to pass the table name. However, in the
code I found two different implementations:
- In Utility class the sequence name should be called as DocumentNo_<tableName>
- In FIN_Utility class the sequence name should be equal to the tablename.
The first implementation seems to be correct according to the triggers, and the second one seems to be a
wrong implementation. In any case, I keep it as I don't want to break the API (although most likely nobody is
using it). Added also a warn message when the case is detected.

Other code review improvements:
- Improved utility class in automated tests
- Added more tests (and more needed to have a better coverage)
- Bug: remove logic that includes the organization's search key in the sequence generation. This is a prefix
as explained in the FS.
- The getDocumentNo() doesn't include now the < >, so it is similar API as the rest of the documentNo() methods.
The caller is in charge of adding them if necessary
- Usage of standard libraries instead of ad-hoc implementations
- Rename methods to make it clear
- Usage of enum to define all the behavior in Java
- Method getDocumentNo() is now recursive to support many levels of sequence calculations