[HU Type Sequence] Deprecate AutoNumbering = N
When this flag was false, the sequence engine was trying to find another sequence
with the name equal to DocumentNo_<tableName>, being the tableName the name of the
table linked to the related Document Type. Therefore all the documents within the
client share the same sequence.

This behavior is extremely weird, and it can be perfectly achieved by defining the
same sequence to the document types of all the client for the same table. Therefore
it has been decided to deprecate it.

A new calculation method called "System DocumentNo_<table> (Deprecated!)" has been
created, and deactivated by default. The purpose is to stop allowing to define this
behavior for the new sequences, but keeping the existing features as they are, so
we don't break this behavior for the existing users.

The Calculation Method is now mandatory field, and the user can only select:
- Auto Numbering (default)
- Based on Sequence

Note that the triggers AD_CLIENT_TRG and AD_COLUMN_TRG2, which automatically populates
the DocumentNo_<tableName> sequences, are kept and not deleted. The reason why is that
there is a hidden feature in columns named as 'VALUE', which can be declared as
IsUsedSequence = Y. In this case the DocumentNo_<tableName> associated will be used to
automatically populate the VALUE column.

Finally, take into account that the AD_CLIENT_TRG populates that info for the tables
that are currently available when the Initial Client Setup is executed. If later on a
new table is created (for example after installing a new module, or because an update
brings a new table) the DocumentNo_<tableName> won't be created automatically. In this
case a module script (or a manual process) might be needed to create the sequences.