[purchaseAssortment] Code review improvements
Avoid creating manual HTML/XML code, and instead using the Standard UI
for Report and Processes.

This requires to create a new column (instead of reusing the existing DocAction)
to avoid issues with WAD.

The old Report and Process linked to M_Requisition_Post has been deprecated, and
the related fields hidden.

The M_REQUISITION_POST now sets the value of new column, so the button in the UI
shows the right label in each moment.