Fixes ISSUE-54368: A new ant task export.translation is required to call to the export translation tool (TranslationManager.exportTrl).
The new task is called by the org.openbravo.translator.helper/src/org/openbravo/translator/helper/scripts/publish.sh script.

The reason for this is to work properly with reduced translations. If the zip file containing the POs was created with the 'Reduced translation' option, all the 'not reduced' fields such as helps and descriptions will be lost. To avoid this, the script will run the full export just after the apply.module (via a new ant task export.translation), to make sure the new translations are in the database, and it will copy the files from the attachments folder to each translated module.