Fixes ISSUE-54685: Remove unnecessary obkmo_widget_class_access rows.
Widget classes marked with allow_anonymous_access = 'Y' are made
available for all roles.
So we do not need to keep obkmo_widget_class_access entries for those
cases.
This is follow-up and very similar to cleanup done in 53408 (for
automatic ad_role  entries).
Creation of such _access rows was already moved from the
obkmo_widget_class_trg in issue 53408.