[nestedhu] Handling Unit | Content filter to hide nested handling units content
M_REFINVENTORY_NESTED function returns a comma-list of nested handling unit for
the given one, including it.

The 'unnest' function is registered in Hibernate to be able to extract this output
and embed it in an HQL query IN or ANY clauses.
Note that it is only declared for PostgreSQL as we don't support Oracle anymore.

The Handling Unit | Content tab HQL where clause is changed to show also the nested
handling units content. Note that the query is doing a dummy subquery inside the ANY
clause to get the same ID as the parameter. This is the only possible way I found to
call the unnest(m_refinventory_nested()).
I tried without the select, or with a select from dual and it didn't work.
The performance of the workaround should be good enough as it uses an Index Only Scan.

The tab has disabled the parent key property to be able to filter by the nested RIs when
the new filter 'Exclude content in nested handling units' is removed