[nestedhu] M_REFINVENTORY_NESTED supports stock with attributesetinstance
For stock already having attributesetinstance, the M_REFINVENTORY_NESTED didn't
work fine, as the innermost attribusetinstance doesn't have a parent_attributesetinstance_id
null or equal to '0'.

The recursive CTE now has a level, so the max level is the innermost attributesetinstance.

The function now receives an attributesetinstance, which is more convinient as it is related
to the storage detail itself.

IMPORTANT: this function won't escalate, because it requires a seq scan to know the storage
details within a referenced inventory. We need to find a better alternative