[nestedhu] Added M_RefInventory.PARENT_REFINVENTORY_ID to allow tree in RI
The tree that we have in M_AttributeSetInstance table might allow to get the RI tree too.
However, this approach has some important problems:
1. If the RI is reused, the M_AttributeSetInstance will have the old RI tree too, and this
might create a big mess. Remember the M_AttributeSetInstance is not modified after unboxing
to keep traceability.

2. It doesn't allow to easily get the nested stock belonging to a RI (and might return weird
things in the RI is being reused). The query to get this info would be very complex and poor
in performance.

Having a tree in RI is nice because:
1. We can easily show as a tree the Handling Unit window, which adds a lot of value for the user.

2. It allows to quickly get the nested RIs. And, since M_StorageDetail is linked to the innermost RI,
it is easier and faster to return all the nested stock in a RI tree.
Note that a PL function with a recursive CTE is pending to be developed.
This will also fix the performance problems in Handling Unit | Content tab when removing the filter.

FIXME: the Parent RI field should be shown as readonly in status bar.