[nestedhu] M_REFINVENTORY_NESTED works with Ref Inventory tree
The M_REFINVENTORY_NESTED function now looks only to the ref inventory tree to
get the nested handling units. The purpose of the function is to return all the
nested RIs, so the p_innermost_only parameter has been removed.
This is now more efficent and safer than previous implementation based on
attributesetinstance tree.

The previous implementation based on the attribute set instance tree is embeded
into the M_UPDATE_INVENTORY, because only this procedure is interested in taking
the innermost RI for the given attribute set instance id.

The Handling Unit | Content tab now uses the new M_REFINVENTORY_NESTED implementation.
However, in my system this doesn't perform properly, because a seq. scan over
m_storage_detail is still performed. It seems the DB is unable to properly use the
index on M_RefInventory_ID column.