[HU Type Sequence] Improvements in ADSequenceEventHandler
Simplified the code assuming this approach: each time a sequence is saved/updated,
I need to validate the MOD10 configuration only if there is any base sequence in the
child tree with alphanumeric prefix/suffix.
In this case I check recursively that this sequence or any other sequence whose base parent
is this sequence is not configured to work with MOD10 control digit.

NEW: in case of consuming the sequence (usual scenario), we skip any configuration validation
to improve performance.

Improved error messages adding parameters, and remove the hint (which is used by translators)
Adapted JUnit tests to use the new messages