Fixes ISSUE-54878: Tab is opened even after hiding translations
The problem is that the application was not supporting hiding a tab, when in the stored version of the view
that tab was active. Now we detect this case, and in the case where the tab that used to be active is no
longer available and there are no more subtabs, updateSubtabVisibility just does not do anything.