Fixes ISSUE-0055012: Problem in loading the data in BUsiness API Data Load window
Problem in loading data through the Business API Data Load window. An InvalidCSRFToken error message is showed in the log.

The problem is that now we are checking that the csrfToken parameter of POST requests, but we missed to include the CSRF token in one of them. So we include it.