Fixes ISSUE-54655: Multiselector now always uses the most updated context
The problem was that when the multiselector window was open, it first invalidated its data using this.selectorWindow.selectorGrid.invalidate cache, which made a request to the datasource
that clashed with the request done when invoking this.selectorWindow.open() a few lines later. The first request did not use the most updated context, and was preventing the second one from
being executed.

To fix this, now data is reset by executing delete this.selectorWindow.selectorGrid.data, that avoids the side effect that was causing the issue