Fixes ISSUE-55134: Event handlers get accurate info on the number of active records
The problem was that when enabling/disabling recordsa using the EnableOrDisableMultipleRecords process, the
event handlers were not receiving accurate info on the number of currently active records if data was flushed
updated updating them all.

In order to prevent this problem, now data is flushed after updating each records. This results in a slower execution
of the process but it is still reasonable fast (changing the active flag of 100 records took <100ms)