Fixes ISSUE-55347: Error when importing Inventory via CSV file in Business API Data Load
The problem here first was a problem of the maximum file count limit has been exceeded during the parsing of the multipart request. For this first we just increase the NUMBER_OF_FIXED_PARAMS to 5 and after we had a second problem related to csrfToken that is invalid, when we search we found that this parameter was added to formData instead of paramValues because we get the csrfToken from those.

So we removed all the new changes related to first phase and we keep just the second ones.