fixes BUG-55640: on change function not executed on newly created records of P&E grids
  The blur property of the P&E grid items was being properly adapted to
execute the on change functions, the first time the record inline editor
was shown.

  But for the case of the grids that accepts creating new records, the
blur function was not being set for the grid items that may be created
whenever a new record is added.

  To fix this problem, now we are always setting the adapted blur
function for all the grid items that do not have it already set
regardless if it is the first time the record inline editor is being
opened or not.