Fixes ISSUE-55100: Callouts are wrongly invoked after refreshing a record
The problem was that, sometimes, due to a race condition, a null value was set to c_bpartner_id after a refresh, that resulted in a
CHANGE request to the FIC that changed some values. Right after, the original c_bpartner_id value was reset, but the values modified
by the FIC remained modified.

The race conditions happens in some part of smartclient code that we were not able to debug (when smartclient was deployed in dev mode,
the issue could not be reproduced in the same way). The trigger for the change done by smartclient was this line [1], where the focus is
removed from a form item when a record is open in form view (or refreshed while in form view), to prevent a double-focus problem in firefox
(see [2]). After removing that line, the problem described in 55100 is no longer reproducible, and the problem described in [2] is not
reproducible either.

A better fix would have been to fix the root cause in smartclient (a form item being blurred should not result in the value of another item
being emptied), but given that it has not been possible to remove the root cause, the current fix seems to be good enough.

[1] https://gitlab.com/openbravo/product/openbravo/-/blob/d54fc216f43a800493d4bc58488ab9820a30d155/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/form/ob-view-form.js\#L285
[2] https://issues.openbravo.com/view.php\?id\=19248