Fixes ISSUE-55963: Avoid double dal-start when executing ant run.test
DAL was initialized twice due to the way OBBaseTest was initializing the
DAL Layer. It was first initialized empty in a @BeforeClass in
OBBaseTest and afterwards it was re-initialized with sqlFunctions in
WeldBaseTest @Before function.

To avoid this double initialization, in WeldBaseTest the @BeforeClass
from OBBaseTest is being hidden by implementing it with the same name
and only triggernig the code that doesn't do initialization of DAL.

DAL initialization is directly handled by the WeldBaseTest, resulting in
only one initialization in those cases. OBBaseTest works as before, if
someone extends it, it will still initialize DAL as before, it only
changes behaviour on tests that extend WeldBaseTest.