Fixes ISSUE-55683: Do not compute tab display logic of ancestors of direct record
When opening a direct record (i.e. by clicking on the header of a field in form view, or by cliking on Recent Records from the workspace,
the tab display logic of the ancestor tabs of the direct record should not be computed, because we need to ensure that the target record
is displayed.

The tab display logic of records that belong to tabs that are not ancestors of the direct record should be computed as usual