fixes BUG-56493: Items of P&E list fields cannot be dynamically changed
  It was not possible to filter the items of a P&E list field through a
FIC call, for example by using a validation or a FIC extension.

  This is because after the FIC response the editor value map is not
updated but just the field value map.

  To fix this problem when are overriding the getEditorValueMap function
of the P&E grids, to return the updated value map from the list field.
This is the same solution that is currently being done by the grids of
the back office standard windows.