Fixes ISSUE-56504: Expired session is never closed
When using AWO and then logging in into openbravo on the same
domain/url, the AWO session was left as inactive but still part of the
context list, however the context list was not actually being checked
for session destroyal. Now sessions are destroyed if the session is
present in the context but are no longer active, because that means
they're created on the same cluster node, and not on another node.